requirements.txt
python-telegram-bot[ext]==21.6
web3>=6.15.0
aiosqlite>=0.19.0
cryptography>=42.0.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
aiofiles>=23.2.0
